Adicionando fornt end no mesmo repositório

+++ .docker/entrypoint.sh
@@ -1,8 +1,19 @@
 #!/bin/bash
 
 #On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
-cp .env.example .env
-cp .env.testing.example .env.testing
+
+### FRONT-END
+npm config set cache /var/www/frontend/.npm-cache --global
+cd frontend && npm install && cd ..
+
+### BRACK-END
+cd backend
+if [ ! -f ".env" ]; then
+  cp .env.example .env
+fi
+if [ ! -f ".env.testing" ]; then
+  cp .env.testing.example .env.testing
+fi
 chown -R www-data:www-data .
 composer install
 php artisan key:generate
